+++ src/main/java/com/logistic/domain/PagamentoComCartao.java
@@ -2,9 +2,11 @@ package com.logistic.domain;
 
 import javax.persistence.Entity;
 
+import com.fasterxml.jackson.annotation.JsonTypeName;
 import com.logistic.domain.enums.EstadoPagamento;
 
 @Entity
+@JsonTypeName("pagamentoComCartao") //tipo se ser indentificado pelo json (@type)
 public class PagamentoComCartao extends Pagamento{
 	
 	public PagamentoComCartao() {}