+++ src/main/java/com/logistic/services/IBoletoService.java
@@ -1,0 +1,11 @@
+package com.logistic.services;
+
+import java.util.Date;
+
+import com.logistic.domain.PagamentoComBoleto;
+
+public interface IBoletoService {
+	
+	void preencherPagamentoComBoleto(PagamentoComBoleto pagto, Date instanteDoPedido);
+
+}