+++ src/main/java/com/logistic/services/generic/IDeleteService.java
@@ -1,0 +1,10 @@
+package com.logistic.services.generic;
+
+import org.springframework.transaction.annotation.Transactional;
+
+public interface IDeleteService<T> extends IService<T> {
+	
+	@Transactional
+	void delete(Integer id);
+	
+}