+++ src/main/java/com/logistic/services/generic/IFindService.java
@@ -5,15 +5,10 @@ import java.util.List;
 import org.springframework.transaction.annotation.Transactional;
 
 @Transactional(readOnly = true)
-public interface IGenericService<T> {
+public interface IFindService<T> extends IService<T> {
 	
 	List<T> findAll();
 	
 	T find(Integer id);
 	
-	T insert(T obj);
-	
-	T update(T obj);
-	
-	void delete(Integer id);
 }