+++ src/main/java/com/logistic/services/generic/IUpdateService.java
@@ -1,0 +1,10 @@
+package com.logistic.services.generic;
+
+import org.springframework.transaction.annotation.Transactional;
+
+public interface IUpdateService<T> extends IService<T> {
+	
+	@Transactional
+	T update(T obj);
+	
+}