+++ src/test/java/com/logistic/services/imp/CategoriaServiceTest.java
@@ -1,5 +1,6 @@
 package com.logistic.services.imp;
 
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.when;
 
 import java.util.ArrayList;
@@ -7,6 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
@@ -14,9 +16,6 @@ import org.mockito.MockitoAnnotations;
 import com.logistic.domain.Categoria;
 import com.logistic.repositories.CategoriaRepository;
 
-// Usar Spring ?? assim ele roda toda a aplicação
-//@RunWith(SpringRunner.class)
-//@SpringBootTest
 public class CategoriaServiceTest {
 	
 	@InjectMocks
@@ -41,19 +40,25 @@ public class CategoriaServiceTest {
 		when(categoriaRepository.findAll()).thenReturn(categoriaList);
 	}
 	
-//	@Test
-//	public void buscarCategoriaPeloId () {
-//		assertThat(categoriaService.find(ID)).isNotNull();
-//	}
-//	
-//	@Test
-//	public void buscarListaCategorias () {
-//		assertThat(categoriaService.findAll()).isNotNull();
-//	}
-
-//	@Test
-//	public void atuaçizarCategoria () {
-//		assertThat(categoriaService.update(categoria)).isNotNull();
-//	}
+	@Test
+	public void buscarCategoriaPeloId () {
+		assertNotNull(categoriaService.find(ID));
+	}
 	
+	@Test
+	public void buscarCategoriaPeloIdQueNaoExiste () {
+		when(categoriaRepository.findById(ID)).thenReturn(Optional.ofNullable(null));
+		try {
+			assertNotNull(categoriaService.find(ID));
+		} catch (Exception e) {
+			//assertThat(e.getMessage(), CoreMatchers.is("Objeto não encontrado! Id: 1, Tipo: com.logistic.domain.Categoria"));
+			//assertThat
+		}
+	}
+	
+	@Test
+	public void buscarListaCategorias () {
+		assertNotNull(categoriaService.findAll());
+	}
+
 }